Fix no public endpoint deployed
