+++ ow_client/src/config/MyWellDevConfig.ts
@@ -47,6 +47,8 @@ const MyWellDevConfig: RemoteConfig = {
     country: { id: 'country', required: true, order: 0 },
   },
   shouldUseV1Search: false,
+  resourceDetail_allowDownload: false,
+  readingDownloadUrl: "https://us-central1-our-water-dev.cloudfunctions.net/public/mywell/downloadReadings",
 }
 
 export default MyWellDevConfig;``
