+++ ow_firebase/functions/lib/src/handlers/fn_public/public.js
@@ -15,6 +15,9 @@ const ErrorHandler_1 = require("../../common/ErrorHandler");
 const QRCode_1 = require("../../common/apis/QRCode");
 const utils_1 = require("../../common/utils");
 const AppProviderTypes_1 = require("ow_common/lib/utils/AppProviderTypes");
+const api_1 = require("ow_common/lib/api");
+const moment = require("moment");
+const FirebaseAdmin_1 = require("../../common/apis/FirebaseAdmin");
 const bodyParser = require('body-parser');
 const Joi = require('joi');
 const fb = require('firebase-admin');
@@ -56,6 +59,43 @@ module.exports = (functions) => {
         const base64Data = qrResult.result.replace(/^data:image\/png;base64,/, "");
         const file = `/tmp/qr_${id}.png`;
         yield utils_1.writeFileAsync(file, base64Data, 'base64');
+        res.download(file);
+    }));
+    /**
+     * Download Readings for Resources
+     * GET /:orgId/downloadReadings
+     *
+     * eg: /test_12348/downloadReadings?resourceIds=00001%2C00002%2C00003
+     *
+     * Download a list of readings for a given resource.
+     * resourceIds must be a comma separated list of resourceIds
+     */
+    const getReadingsValidation = {
+        options: {
+            allowUnknownBody: false,
+        },
+        query: {
+            resourceIds: Joi.string().required(),
+        }
+    };
+    app.get('/:orgId/downloadReadings', validate(getReadingsValidation), (req, res) => __awaiter(this, void 0, void 0, function* () {
+        let { resourceIds } = req.query;
+        console.log("Getting readings:", resourceIds);
+        const { orgId } = req.params;
+        const readingApi = new api_1.ReadingApi(FirebaseAdmin_1.firestore, orgId);
+        resourceIds = resourceIds.split(',');
+        if (resourceIds.length > 50) {
+            throw new Error("Too many resourceIds. Max is 50");
+        }
+        const readings = AppProviderTypes_1.unsafeUnwrap(yield readingApi.getReadingsForResources(resourceIds, { limit: 100 }));
+        if (readings.readings.length === 0) {
+            const error = new Error(`No readings found for ids: ${resourceIds}`);
+            return res.status(404).send(error);
+        }
+        const readingsData = api_1.ExportApi.readingsToExport(readings.readings, api_1.ExportFormat.CSV);
+        console.log("readings data is", readingsData);
+        const file = `/tmp/${moment.toString()}.csv`;
+        yield utils_1.writeFileAsync(file, readingsData, 'utf-8');
         res.download(file);
     }));
     /* CORS Configuration */
