+++ ow_firebase/functions/src/handlers/fn_admin/admin.ts
@@ -4,14 +4,13 @@ import * as cors from 'cors';
 import ErrorHandler from '../../common/ErrorHandler';
 
 import { validateFirebaseIdToken, validateUserIsAdmin } from '../../middleware';
-import { enableLogging, writeFileAsync } from '../../common/utils';
+import { enableLogging } from '../../common/utils';
 import FirebaseApi from '../../common/apis/FirebaseApi';
 import { firestore } from '../../common/apis/FirebaseAdmin';
-import { ResultType, unsafeUnwrap } from 'ow_common/lib/utils/AppProviderTypes';
-import { UserApi, ReadingApi, ExportApi, ExportFormat } from 'ow_common/lib/api';
+import { ResultType } from 'ow_common/lib/utils/AppProviderTypes';
+import { UserApi } from 'ow_common/lib/api';
 import UserType from 'ow_common/lib/enums/UserType';
 import UserStatus from 'ow_common/lib/enums/UserStatus';
-import * as moment from 'moment';
 
 
 const bodyParser = require('body-parser');
@@ -98,49 +97,6 @@ module.exports = (functions) => {
     }
 
     res.status(204).send("true");
-  });
-
-  /**
-   * Download Readings for Resources
-   * GET /:orgId/downloadReadings
-   * 
-   * eg: /test_12348/downloadReadings?resourceIds=00001%2C00002%2C00003
-   * 
-   * Download a list of readings for a given resource. 
-   * resourceIds must be a comma separated list of resourceIds
-   */
-  const getReadingsValidation = {
-    options: {
-      allowUnknownBody: false,
-    },
-    query: {
-      resourceIds: Joi.string().required(),
-    }
-  }
-  app.get('/:orgId/downloadReadings', validate(getReadingsValidation), async (req, res) => {
-    
-    let { resourceIds } = req.query;
-    console.log("Getting readings:", resourceIds);
-    const { orgId } = req.params;
-    const readingApi = new ReadingApi (firestore, orgId);
-
-    resourceIds = resourceIds.split(',');
-    if (resourceIds.length > 50) {
-      throw new Error("Too many resourceIds. Max is 50");
-    }
-
-    const readings = unsafeUnwrap(await readingApi.getReadingsForResources(resourceIds, {limit: 100}));
-    if (readings.readings.length === 0) {
-      const error = new Error(`No readings found for ids: ${resourceIds}`);
-      return res.status(404).send(error);
-    }
-    const readingsData = ExportApi.readingsToExport(readings.readings, ExportFormat.CSV);
-    console.log("readings data is", readingsData);
-
-    const file = `/tmp/${moment().toString()}.csv`;
-    await writeFileAsync(file, readingsData, 'utf-8');
-    
-    res.download(file);
   });
 
   /* CORS Configuration */
