+++ ow_firebase/functions/src/handlers/fn_public/public.ts
@@ -2,15 +2,15 @@ import * as validate from 'express-validation';
 import * as express from 'express';
 import * as cors from 'cors';
 import ErrorHandler from '../../common/ErrorHandler';
-
-
-//@ts-ignore
-import * as morgan from 'morgan';
-//@ts-ignore
-import * as morganBody from 'morgan-body';
 import { generateQRCode } from '../../common/apis/QRCode';
 import { writeFileAsync, enableLogging } from '../../common/utils';
-import { ResultType } from 'ow_common/lib/utils/AppProviderTypes';
+import { ResultType, unsafeUnwrap } from 'ow_common/lib/utils/AppProviderTypes';
+import { ReadingApi, ExportApi, ExportFormat } from 'ow_common/lib/api';
+import * as moment from 'moment';
+import { firestore } from '../../common/apis/FirebaseAdmin';
+
+
+
 
 const bodyParser = require('body-parser');
 const Joi = require('joi');
@@ -66,6 +66,50 @@ module.exports = (functions) => {
 
     res.download(file);
   });
+
+  /**
+   * Download Readings for Resources
+   * GET /:orgId/downloadReadings
+   * 
+   * eg: /test_12348/downloadReadings?resourceIds=00001%2C00002%2C00003
+   * 
+   * Download a list of readings for a given resource. 
+   * resourceIds must be a comma separated list of resourceIds
+   */
+  const getReadingsValidation = {
+    options: {
+      allowUnknownBody: false,
+    },
+    query: {
+      resourceIds: Joi.string().required(),
+    }
+  }
+  app.get('/:orgId/downloadReadings', validate(getReadingsValidation), async (req, res) => {
+
+    let { resourceIds } = req.query;
+    console.log("Getting readings:", resourceIds);
+    const { orgId } = req.params;
+    const readingApi = new ReadingApi(firestore, orgId);
+
+    resourceIds = resourceIds.split(',');
+    if (resourceIds.length > 50) {
+      throw new Error("Too many resourceIds. Max is 50");
+    }
+
+    const readings = unsafeUnwrap(await readingApi.getReadingsForResources(resourceIds, { limit: 100 }));
+    if (readings.readings.length === 0) {
+      const error = new Error(`No readings found for ids: ${resourceIds}`);
+      return res.status(404).send(error);
+    }
+    const readingsData = ExportApi.readingsToExport(readings.readings, ExportFormat.CSV);
+    console.log("readings data is", readingsData);
+
+    const file = `/tmp/${moment.toString()}.csv`;
+    await writeFileAsync(file, readingsData, 'utf-8');
+
+    res.download(file);
+  });
+
 
   /* CORS Configuration */
   const openCors = cors({ origin: '*' });
