+++ ow_firebase/functions/src/index.js
@@ -60,6 +60,11 @@ if (!process.env.FUNCTION_NAME || process.env.FUNCTION_NAME === 'sync') {
   exports.sync = require('./handlers/fn_sync/sync')(functions);
 }
 
+//Sync Api
+if (!process.env.FUNCTION_NAME || process.env.FUNCTION_NAME === 'public') {
+  exports.public = require('./handlers/fn_public/public')(functions);
+}
+
 
 
 /**
